Delete Product API

+++ src/components/Dashboard/AdminProductRow.js
@@ -6,20 +6,33 @@ import {
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import React from "react";
 import toast from "react-hot-toast";
+import Swal from "sweetalert2";
 
 const AdminProductRow = ({ product, index, refetch }) => {
   const handleDel = (id) => {
-    fetch(`http://localhost:5000/products/${id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        console.log(data);
-        if (data.deletedCount) {
-          toast.success("Product Deleted");
-          refetch();
-        }
-      });
+    Swal.fire({
+      title: "Are you sure?",
+      text: `You are deleting orderID: ${product.productName}. You can't revert it.`,
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`http://localhost:5000/products/${id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            console.log(data);
+            if (data.deletedCount) {
+              Swal.fire("Deleted!", "The Product has been deleted.", "success");
+              refetch();
+            }
+          });
+      }
+    });
   };
   return (
     <tr>
@@ -35,7 +48,7 @@ const AdminProductRow = ({ product, index, refetch }) => {
       <td>{product.description.slice(0, 10)}</td>
       <td className="flex gap-1">
         <FontAwesomeIcon
-        onClick={()=>handleDel(product._id)}
+          onClick={() => handleDel(product._id)}
           className="font-2xl cursor-pointer bg-red-600 p-2 rounded text-white"
           icon={faTrashCan}
         ></FontAwesomeIcon>
